automatically backed up by learn

+++ lib/bachelor.rb
@@ -44,9 +44,9 @@ def get_average_age_for_season(data, season)
   sum = 0
   divisor = data[season].length
   data[season].each do |contestants_data|
-    sum += contestants_data["age"].to_i
+    sum += contestants_data["age"].to_f
   end
-  return (sum / divisor).to_f
+  return (sum / divisor).to_i
 end
